Add listMapper and tests.
The list mapper handles all interaction between the List
object and it's database communication. In a perfect world
it will be used internally by List and will not appear anywhere
else.

This version is a synchronous API with return values. I actually *need*
to change this to make it async. Otherwise A: It won't work and B: It
won't be good Node-y style.